refactor(Core): #394 redirect Crawler-targeting tests + slim CoreTests deps
Twelfth leaf in the DI epic. Core's *production* deps were already
clean (CoreProtocols, CoreJSONParser, CorePackageIndexing, Shared*,
Logging, Resources, ASTIndexer — all foundation-tier and DI-leaf-
pinned). The DI violation lived in CoreTests, which depended on
Crawler + Search because the test files that exercised them were
historically housed under Tests/CoreTests/ — a relic of the v1.0.x
era when Crawler lived inside the Core target.

Changes:

1. Move three dedicated Crawler test files from Tests/CoreTests/ to
   Tests/CrawlerTests/:
   - CrawlerTests.swift (TechnologiesIndex fetch + crawler engine
     fixtures)
   - ArchiveGuideCatalogTests.swift (Apple Archive guide catalog)
   - SwiftEvolutionCrawlerTests.swift (Swift Evolution crawler)

2. Extract every Crawler.AppleDocs / Crawler.AppleDocs.State test
   that was buried in CupertinoCoreTests (~630 lines: 1
   integration test + 13 State change-detection tests + the
   private test helpers createTempDirectory /
   createTestConfiguration / verifyOutputDirectory /
   verifyMarkdownFiles / etc.) into a new file
   Tests/CrawlerTests/Crawler.AppleDocs.IntegrationTests.swift.
   No behavioural change; this is a pure relocation.

3. Drop the now-dead imports from CupertinoCoreTests:
   - import Crawler (no remaining call sites)
   - @testable import Search (was already dead — line 78's
     "swiftPackagesCatalogSearch" is a function name, not Search-
     the-target; line 80's "Search for" is text in a string).

4. Slim CoreTests target deps in Package.swift:
   - Drop "Crawler" (no test in Tests/CoreTests/ imports it now)
   - Drop "Search"   (no test in Tests/CoreTests/ imports it now)

   Remaining CoreTests deps:
     ["CoreProtocols", "CoreJSONParser", "CorePackageIndexing",
      "Core", "SharedCore", "SharedConstants", "SharedModels",
      "TestSupport"]
   — every entry matches a production import in the Core target
   or supports a Core-local test fixture.

The CrawlerTests target deps already include Core (Crawler depends
on Core in production, so its tests inherit that), so the moved
files compile without further changes.

Verification: xcrun swift build clean; make test-clean -> 1404
tests in 156 suites pass (was 1404, no regressions; the moved
tests run from their new home).

Closes #394
